feat: Cleaned up PR for review
