+++ src/components/search/components/searchFiltersMenu/components/SelectFilter.tsx
@@ -16,6 +16,9 @@ import GetDigitalSpecimenSearchTermAggregations from "api/digitalSpecimen/GetDig
 /* Import Components */
 import { MultiSelect } from 'components/elements/customUI/CustomUI';
 
+/* Import utilities */
+import { formatMissingDataFilter } from 'app/utilities/SearchFilterUtilities';
+
 
 /* Props Type */
 type Props = {
@@ -56,7 +59,7 @@ const SelectFilter = (props: Props) => {
     fieldValues.forEach(fieldValue => {
         /* Prepend selected item */
         multiSelectItems.unshift({
-            label: isMissingDataFilter ? fieldValue.replace('no', 'No ') : fieldValue,
+            label: isMissingDataFilter ? formatMissingDataFilter(fieldValue, 'title') : fieldValue,
             value: fieldValue,
             count: filterAggregations?.[fieldValue] ?? aggregations?.[fieldValue] ?? bootAggregations?.[name]?.[fieldValue]
         });
@@ -68,7 +71,7 @@ const SelectFilter = (props: Props) => {
             /* Check if item is not already present due to selected list */
             if (!multiSelectItems.find(multiSelectItem => multiSelectItem.value === key)) {
                 multiSelectItems.push({
-                    label: isMissingDataFilter ? key.replace('no', 'No ') : key,
+                    label: isMissingDataFilter ? formatMissingDataFilter(key, 'title') : key,
                     value: key,
                     count
                 });
@@ -83,7 +86,7 @@ const SelectFilter = (props: Props) => {
             /* Check if item is not already present due to selected list */
             if (!multiSelectItems.find(multiSelectItem => multiSelectItem.value === key)) {
                 multiSelectItems.push({
-                    label: isMissingDataFilter ? key.replace('no', 'No ') : key,
+                    label: isMissingDataFilter ? formatMissingDataFilter(key, 'title') : key,
                     value: key,
                     count
                 });
